Créer une fonction qui récupère les donneés d'un dossier et renvoie les balises formattées pour la génération de documents
